Add delay buffer allocation limit
Revert "Increase delay buffer counter width" and instead make the server
abort() when reaching the limit.

+++ common/player.cpp
@@ -47,7 +47,7 @@ void clear_player(player *pl) {
     if (pl->unarmed_skill)
         FREE_AND_CLEAR_STR(pl->unarmed_skill);
 
-    for (uint16_t db = 0; db < pl->delayed_buffers_allocated; db++) {
+    for (uint8_t db = 0; db < pl->delayed_buffers_allocated; db++) {
         free(pl->delayed_buffers[db]);
     }
     FREE_AND_CLEAR(pl->delayed_buffers);
